Add files via upload
This report is about our mini project, which in called as Envision Plus . Developed a cutting-edge pair of smart glasses utilizing Raspberry
Pi and Arduino Uno to enhance daily life for visually impaired users. Integrated
obstacle detection technology to alert users of nearby objects and a text-to-speech
converter to read out text and providing real-time audio feedback. This project
showcases a commitment to leveraging technology for accessibility and practical
problem-solving.